+++ packages/server/drizzle.config.ts
@@ -1,12 +1,11 @@
 import { defineConfig } from "drizzle-kit";
-import { join } from "path";
+import { ENV } from "./src/lib";
 
 export default defineConfig({
     schema: './src/db/schema.ts',
     out: './drizzle',
     dialect: 'sqlite',
     dbCredentials: {
-        url: (process.env.DB_PATH && process.env.DB_FILE) ?
-            join(process.env.DB_PATH, process.env.DB_FILE) : 'plant.db'
+        url: ENV.DATABASE_PATH
     }
 })